Stop implying that stack-allocating events is a good performance hack.
Sticking the struct event inside a struct can shave off a nanosecond
here or there, especially if you're seeing cache misses, and it can
save a few bytes per event in malloc overhead.  But putting the event
on the stack is not a great trick; if you use enough events for the
malloc overhead to matter, that's more RAM than you should be putting
on the stack.