Don't contradict referencing sentence in footnote
The footnote claiming that reading the fd_set takes time proportional to
the number of fds is confusing, since both the sentence referencing that
footnote and the example right above that sentence are saying that the
time is proportional to the largest fd.

Of course it's possible to do better by storing the fds separately, but
the footnote should not imply that this was done here.